Update file paths
