+++ src/App.css
@@ -1,113 +1,141 @@
 body {
-  font-family: 'Inter', 
-               system-ui, 
-               -apple-system, 
-               BlinkMacSystemFont, 
-               'Segoe UI', 
-               Roboto, 
-               Ubuntu, 
-               'Helvetica Neue', 
-               sans-serif;
+  background: white !important;
+  font-family: 'Inter';
+  color: '#666666' !important;
+}
+ main {
+  background: white !important;
+ }
+
+footer {
+  background: #F8F8F8 !important;
 }
 
-.unauthorized {
+#login {
+  background-color: #00c08b;
+  border-color: #00c08b;
+  border-radius: 5px;
+}
+
+.btn-custom:disabled {
+  color: #666666;
+  background-color: #e5e5e5;
+  border-color: #e5e5e5;
+  border-radius: 5px;
+}
+
+.list-item {
+  padding: 16px;
+  margin: 0;
+}
+
+.flex {
   display: flex;
+}
+
+.flex-center {
   justify-content: center;
+}
+
+.flex-start {
+  justify-content: flex-start;
+}
+
+.flex-end {
+  justify-content: flex-end;
+}
+
+.space-between {
+  justify-content: space-between;
+}
+
+.align-center {
   align-items: center;
-  padding: 0;
 }
 
 .logo {
-  display: block;
-  height: 60px;
+  height: 70px;
 }
 
-#top-nav {
+.separator {
+  padding: 0 10px;
+}
+
+.desktop {
+  max-width: 1152px;
+}
+
+.header {
   display: flex;
-  justify-content: space-between;
+  padding: 50px;
+  margin: 0;
+  width: 100%;
+  min-height: 100px;
   background: white;
 }
 
-footer {
-  display: flex;
-  justify-content: center;
+h1 {
+  font-size: 48px;
+  margin: 0;
+  padding: 0;
+  font-weight: 300;
 }
 
-.main {
-  min-height: 100vh;
+h2 {
+  font-size: 30px;
+  margin: 0;
+  padding: 0;
+  min-width: 100%;
+  height: 100%;
+  font-weight: 300;
 }
 
-/* 
-.App-header {
-  background-image: url('https://nearprotocol.com/wp-content/uploads/2019/03/illo-developers.svg');
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: 1rem;
-  color: white;
-  text-align: center;
-  padding:2%; 
+.content { 
+  padding: 0 144px 70px 144px;
 }
 
-.App-link {
-  color: #61dafb;
+h3 {
+  font-size: 24px;
+  margin: 0;
+  padding: 0;
 }
 
-.Transfer-module {
-  color: black;
+.subheader-lg {
+  font-size: 22px;
+  font-weight: bold;
+  margin: 0;
+  padding: 0;
 }
 
-.logo-wrapper {
-  display: flex;
-  flex-direction: row-reverse;
+.subheader-md {
+  font-size: 18px;
+  font-weight: bold;
+  margin: 0;
+  padding: 0;
 }
 
-.App-logo {
-  animation: App-logo-spin infinite 20s linear;
-  pointer-events: none;
+.near-label {
+  font-size: 14px;
+  text-transform: uppercase;
+  font-weight: 500;
+  margin: 0;
+  padding: 0;
 }
 
-.margin-logo {
-  margin: 20px;
+.ant-btn {
+  font-size: 18px;
 }
 
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
+.description {
+  width: 500px;
 }
 
-.image-wrapper {
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-  justify-content: center;
+.text {
+  font-size: 14px;
 }
 
-.logo {
-  width: 33%;
+.ant-list-header {
+  font-size: 48px;
+  font-weight: 300;
+  padding-bottom: 48px;
 }
-
-p {
-  color: #25282A;
-  font-family: sans-serif;
-}
-
-button {
-  padding: 5px;
-  color: #0072CE;
-  border-radius: 10px;
-  border: 3px solid #0072CE;
-}
-
-button:hover {
-  cursor: pointer;
-  color: white;
-  background: #0072CE;
-}
-*/
