+++ src/App.prev.css
@@ -1,0 +1,113 @@
+body {
+  font-family: 'Inter', 
+               system-ui, 
+               -apple-system, 
+               BlinkMacSystemFont, 
+               'Segoe UI', 
+               Roboto, 
+               Ubuntu, 
+               'Helvetica Neue', 
+               sans-serif;
+}
+
+.unauthorized {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 0;
+}
+
+.logo {
+  display: block;
+  height: 60px;
+}
+
+#top-nav {
+  display: flex;
+  justify-content: space-between;
+  background: white;
+}
+
+footer {
+  display: flex;
+  justify-content: center;
+}
+
+.main {
+  min-height: 100vh;
+}
+
+/* 
+.App-header {
+  background-image: url('https://nearprotocol.com/wp-content/uploads/2019/03/illo-developers.svg');
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  font-size: 1rem;
+  color: white;
+  text-align: center;
+  padding:2%; 
+}
+
+.App-link {
+  color: #61dafb;
+}
+
+.Transfer-module {
+  color: black;
+}
+
+.logo-wrapper {
+  display: flex;
+  flex-direction: row-reverse;
+}
+
+.App-logo {
+  animation: App-logo-spin infinite 20s linear;
+  pointer-events: none;
+}
+
+.margin-logo {
+  margin: 20px;
+}
+
+@keyframes App-logo-spin {
+  from {
+    transform: rotate(0deg);
+  }
+  to {
+    transform: rotate(360deg);
+  }
+}
+
+.image-wrapper {
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+  justify-content: center;
+}
+
+.logo {
+  width: 33%;
+}
+
+p {
+  color: #25282A;
+  font-family: sans-serif;
+}
+
+button {
+  padding: 5px;
+  color: #0072CE;
+  border-radius: 10px;
+  border: 3px solid #0072CE;
+}
+
+button:hover {
+  cursor: pointer;
+  color: white;
+  background: #0072CE;
+}
+*/
