--- src/App2.css
@@ -1,141 +1,0 @@
-body {
-  background: white !important;
-  font-family: 'Inter';
-  color: '#666666' !important;
-}
- main {
-  background: white !important;
- }
-
-footer {
-  background: #F8F8F8 !important;
-}
-
-#login {
-  background-color: #00c08b;
-  border-color: #00c08b;
-  border-radius: 5px;
-}
-
-.btn-custom:disabled {
-  color: #666666;
-  background-color: #e5e5e5;
-  border-color: #e5e5e5;
-  border-radius: 5px;
-}
-
-.list-item {
-  padding: 16px;
-  margin: 0;
-}
-
-.flex {
-  display: flex;
-}
-
-.flex-center {
-  justify-content: center;
-}
-
-.flex-start {
-  justify-content: flex-start;
-}
-
-.flex-end {
-  justify-content: flex-end;
-}
-
-.space-between {
-  justify-content: space-between;
-}
-
-.align-center {
-  align-items: center;
-}
-
-.logo {
-  height: 70px;
-}
-
-.separator {
-  padding: 0 10px;
-}
-
-.desktop {
-  max-width: 1152px;
-}
-
-.header {
-  display: flex;
-  padding: 50px;
-  margin: 0;
-  width: 100%;
-  min-height: 100px;
-  background: white;
-}
-
-h1 {
-  font-size: 48px;
-  margin: 0;
-  padding: 0;
-  font-weight: 300;
-}
-
-h2 {
-  font-size: 30px;
-  margin: 0;
-  padding: 0;
-  min-width: 100%;
-  height: 100%;
-  font-weight: 300;
-}
-
-.content { 
-  padding: 0 144px 70px 144px;
-}
-
-h3 {
-  font-size: 24px;
-  margin: 0;
-  padding: 0;
-}
-
-.subheader-lg {
-  font-size: 22px;
-  font-weight: bold;
-  margin: 0;
-  padding: 0;
-}
-
-.subheader-md {
-  font-size: 18px;
-  font-weight: bold;
-  margin: 0;
-  padding: 0;
-}
-
-.near-label {
-  font-size: 14px;
-  text-transform: uppercase;
-  font-weight: 500;
-  margin: 0;
-  padding: 0;
-}
-
-.ant-btn {
-  font-size: 18px;
-}
-
-.description {
-  width: 500px;
-}
-
-.text {
-  font-size: 14px;
-}
-
-.ant-list-header {
-  font-size: 48px;
-  font-weight: 300;
-  padding-bottom: 48px;
-}
